fix all the names and namespaces review comments
