+++ source/Messaging/source/Communication.IntegrationTests/integrationtest.local.settings.sample.json
