+++ source/Messaging/source/Communication.UnitTests/Communication.UnitTests.csproj
@@ -17,16 +17,16 @@ limitations under the License.
 
   <PropertyGroup>
     <IsPackable>false</IsPackable>
-    <AssemblyName>Energinet.DataHub.Core.Messaging.UnitTests</AssemblyName>
-    <RootNamespace>Energinet.DataHub.Core.Messaging.UnitTests</RootNamespace>
+    <AssemblyName>Energinet.DataHub.Core.Messaging.Communication.UnitTests</AssemblyName>
+    <RootNamespace>Energinet.DataHub.Core.Messaging.Communication.UnitTests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="System.Linq.Async" Version="6.0.1" />
-    <PackageReference Include="xunit" Version="2.7.0" />
-    <PackageReference Include="FluentAssertions" Version="6.12.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
-    <PackageReference Include="xunit.categories" Version="2.0.8" />
+    <PackageReference Include="System.Linq.Async" Version="6.0.1"/>
+    <PackageReference Include="xunit" Version="2.7.0"/>
+    <PackageReference Include="FluentAssertions" Version="6.12.0"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0"/>
+    <PackageReference Include="xunit.categories" Version="2.0.8"/>
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.7">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
@@ -38,6 +38,6 @@ limitations under the License.
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Messaging\Messaging.csproj" />
+    <ProjectReference Include="..\Communication\Communication.csproj"/>
   </ItemGroup>
 </Project>
