+++ source/Messaging/source/Communication.UnitTests/IntegrationEventHandlerStub.cs
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Messaging.Communication;
-using Energinet.DataHub.Core.Messaging.Communication.Publisher;
+using Energinet.DataHub.Core.Messaging.Communication.Subscriber;
 
-namespace Energinet.DataHub.Core.Messaging.UnitTests;
+namespace Energinet.DataHub.Core.Messaging.Communication.UnitTests;
 
-public class IntegrationEventProviderStub : IIntegrationEventProvider
+public class IntegrationEventHandlerStub : IIntegrationEventHandler
 {
-    public IAsyncEnumerable<IntegrationEvent> GetAsync()
-        => Array.Empty<IntegrationEvent>().ToAsyncEnumerable();
+    public Task HandleAsync(IntegrationEvent integrationEvent)
+    {
+        return Task.CompletedTask;
+    }
 }
