+++ source/Messaging/source/Communication/IntegrationEvent.cs
