+++ source/Messaging/source/Communication/Internal/Subscriber/IIntegrationEventFactory.cs
