+++ source/Messaging/source/Communication/Publisher/IPublisher.cs
