+++ source/Messaging/source/Communication/Registration.cs
