+++ source/Messaging/source/Communication/Subscriber/ISubscriber.cs
